visitas crud working

+++ src/controllers/index.controller.js
@@ -10,8 +10,8 @@ import {
   getVisitas,
   getClientes,
   deleteVisita,
+  getUsusarioTecnico,
 } from "./db.contoller";
-import { refreshDynamicContent } from "../handler/events.handler";
 
 export function Home() {
   const divElement = document.createElement("div");
@@ -42,6 +42,7 @@ export async function printDataInventario() {
     const inventario_data = await getProducts();
     const empresas_data = await getEmpresas();
     const selectForm = document.getElementById("nit_suministradora");
+    const tablaBody_N = document.getElementById("tablaBody_N");
 
     inventario_data.forEach((producto) => {
       const fila = document.createElement("tr");
@@ -63,59 +64,49 @@ export async function printDataInventario() {
           }"><span></span></button>
         </td>
       `;
+
       tablaBody_N.appendChild(fila);
 
       // Aquí, después de agregar la fila, agregamos el event listener para el botón de eliminar
-
       const btn_delete = fila.querySelector(".btn_delete");
       btn_delete.addEventListener("click", async (e) => {
-        try {
-          // Obtener el ID del producto que se va a eliminar desde el atributo data-product-id del botón
-          const productId = e.target.dataset.productId;
+        if (confirm("¿Estás seguro de que deseas eliminar este producto?")) {
+          try {
+            // Obtener el ID del producto que se va a eliminar desde el atributo data-product-id del botón
+            const productId = e.target.dataset.productId;
+            const deleteP = await deleteProduct(productId);
 
-          const deleteP = await deleteProduct(productId);
+            // Mostrar un mensaje de éxito o actualizar la interfaz de usuario según sea necesario
+            const alertPlaceholder = document.getElementById(
+              "liveAlertPlaceholder"
+            );
+            const appendAlert = (message, type) => {
+              const wrapper = document.createElement("div");
+              wrapper.innerHTML = [
+                `<div class="alert alert-${type} alert-dismissible" role="alert">`,
+                `   <div>${message}</div>`,
+                '   <button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>',
+                "</div>",
+              ].join("");
 
-          // Mostrar un mensaje de éxito o actualizar la interfaz de usuario según sea necesario
+              alertPlaceholder.append(wrapper);
+            };
 
-          const alertPlaceholder = document.getElementById(
-            "liveAlertPlaceholder"
-          );
-          const appendAlert = (message, type) => {
-            const wrapper = document.createElement("div");
-            wrapper.innerHTML = [
-              `<div class="alert alert-${type} alert-dismissible" role="alert">`,
-              `   <div>${message}</div>`,
-              '   <button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>',
-              "</div>",
-            ].join("");
+            if (deleteP) {
+              appendAlert("Producto eliminado exitosamente", "danger");
+            }
 
-            alertPlaceholder.append(wrapper);
-          };
-
-          if (deleteP) {
-            appendAlert("Producto eliminado exitosamente", "danger");
+            printDataInventario();
+          } catch (error) {
+            console.error("Error al eliminar el producto:", error);
+            // Puedes mostrar un mensaje de error al usuario si la eliminación falla
           }
-
-          printDataInventario();
-        } catch (error) {
-          console.error("Error al eliminar el producto:", error);
-          // Puedes mostrar un mensaje de error al usuario si la eliminación falla
         }
       });
     });
 
-    async function printProductByID() {
-      const productId = 1;
-      try {
-        const product = await getProductById(productId);
-        console.log(product);
-      } catch (error) {
-        console.error(error);
-      }
-    }
-
+   
     // este foreach llama los datos de la tabla empresas para mostrarlas en el formulario de inventario y poder elegir un nit ya existente
-
     empresas_data.forEach((empresa) => {
       const optionElement = document.createElement("option");
       optionElement.value = empresa.Nombre_Empresa_Suministradora;
@@ -129,140 +120,20 @@ export async function printDataInventario() {
   }
 }
 
-// ---------------------visitas-------------------------
-
-// export async function printVisitas() {
-//   const contenidoDinamico = document.getElementById("contenidoDinamico");
-//   contenidoDinamico.classList = "fade-in-out";
-//   contenidoDinamico.innerHTML = hVisitasView;
-
-//   try {
-//     const hVisitas_data = await getVisitas();
-//     const clientes_data = await getClientes();
-//     const inventario_data = await getProducts();
-//     const clienteCC_select_input = document.getElementById(
-//       "clienteCC_select_input"
-//     );
-//     const cuentaCliente_select = document.getElementById("cuentaClienteSelect");
-//     const clienteProducto_select = document.getElementById("clienteProducto");
-
-//     hVisitas_data.forEach((visita) => {
-//       const fila = document.createElement("tr");
-//       fila.innerHTML = `
-//       <td>${visita.Pk_Id_Visita_Tecnica}</td>
-//       <td>${new Date(visita.Fecha_Visita_Tecnica).toLocaleDateString()}</td>
-//       <td>${visita.Tipo_Visita}</td>
-
-//       <td>
-//           <button type="button" class="btn btn-info btn-sm" data-bs-toggle="modal" data-bs-target="#descripcionModal_${
-//             visita.id
-//           }">
-//               Ver Descripción
-//           </button>
-//       </td>
-//       <!-- Agrega el modal aquí -->
-//       <div class="modal fade" id="descripcionModal_${
-//         visita.id
-//       }" tabindex="-1" aria-labelledby="descripcionModalLabel_${
-//           visita.id
-//         }" aria-hidden="true">
-//           <div class="modal-dialog">
-//               <div class="modal-content">
-//                   <div class="modal-header">
-//                       <h5 class="modal-title" id="descripcionModalLabel_${
-//                         visita.id
-//                       }">Descripción de la Visita</h5>
-//                       <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-//                   </div>
-//                   <div class="modal-body">
-//                       ${visita.Descripcion_Visita_Tecnica}
-//                   </div>
-//               </div>
-//           </div>
-//       </div>
-
-//       <td>${visita.Fk_CC_Usuario}</td>
-//       <td>${visita.Fk_Id_Producto}</td>
-//       <td>${visita.Fk_Numero_Cuenta_Cliente}<td>
-
-//       <td>
-//             <button type="button" class="btn_edit rounded btn btn-primary" data-product-id="${
-//               visita.Pk_Id_Visita_Tecnica
-//             }"><span></span></button>
-//             <button type="button" class="btn_delete rounded btn btn-danger" data-product-id="${
-//               visita.Pk_Id_Visita_Tecnica
-//             }"><span></span></button>
-//           </td>
-//       `;
-//         tablaBody_V.appendChild(fila);
-
-//       const btn_delete = fila.querySelector(".btn_delete");
-//       btn_delete.addEventListener("click", async (e) => {
-//         try {
-//           const visitaId = e.target.dataset.productId;
-//           const deleteP = await deleteVisita(visitaId);
-
-//           const alertPlaceholder = document.getElementById(
-//             "liveAlertPlaceholder"
-//           );
-//           const appendAlert = (message, type) => {
-//             const wrapper = document.createElement("div");
-//             wrapper.innerHTML = [
-//               `<div class="alert alert-${type} alert-dismissible" role="alert">`,
-//               `   <div>${message}</div>`,
-//               '   <button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>',
-//               "</div>",
-//             ].join("");
-
-//             alertPlaceholder.append(wrapper);
-//           };
-
-//           if (deleteP) {
-//             appendAlert("Visita eliminado exitosamente", "danger");
-//           }
-//           printVisitas()
-//         } catch (error) {
-//           console.error(error);
-//         }
-//       });
-//     });
-
-//     clientes_data.forEach((cliente) => {
-//       const optionElement = document.createElement("option");
-//       optionElement.value = cliente.Cedula_Cliente;
-//       optionElement.textContent = `${cliente.Nombre_Cliente}  ${cliente.Apellido_Cliente} CC: ${cliente.Cedula_Cliente} `;
-//       clienteCC_select_input.appendChild(optionElement);
-//     });
-
-//     clientes_data.forEach((cliente) => {
-//       const optionElement = document.createElement("option");
-//       optionElement.value = cliente.Pk_Numero_Cuenta_Cliente;
-//       optionElement.textContent = `${cliente.Nombre_Cliente}  ${cliente.Apellido_Cliente} CC: ${cliente.Pk_Numero_Cuenta_Cliente} `;
-//       cuentaCliente_select.appendChild(optionElement);
-//     });
-
-//     inventario_data.forEach((product) => {
-//       const optionElement = document.createElement("option");
-//       optionElement.value = product.Pk_Id_Producto;
-//       optionElement.textContent = `${product.Nombre_Producto} ID: ${product.Referencia} `;
-//       clienteProducto_select.appendChild(optionElement);
-//     });
-//   } catch (error) {
-//     console.error(error);
-//   }
-// }
-
 export async function printVisitas() {
   const contenidoDinamico = document.getElementById("contenidoDinamico");
   contenidoDinamico.classList = "fade-in-out";
   contenidoDinamico.innerHTML = hVisitasView;
 
   try {
-    const [hVisitas_data, clientes_data, inventario_data] = await Promise.all([
-      getVisitas(),
-      getClientes(),
-      getProducts(),
-    ]);
+     const [hVisitas_data, clientes_data, inventario_data, tecnicos_data] =
+      await Promise.all([
+        getVisitas(),
+        getClientes(),
+        getProducts(),
+        getUsusarioTecnico(),
+      ]);
+
     const clienteCC_select_input = document.getElementById(
       "clienteCC_select_input"
     );
@@ -273,28 +144,51 @@ export async function printVisitas() {
     hVisitas_data.forEach((visita) => {
       const fila = document.createElement("tr");
       fila.innerHTML = `
-        <td>${visita.Pk_Id_Visita_Tecnica}</td>
-        <td>${new Date(visita.Fecha_Visita_Tecnica).toLocaleDateString()}</td>
-        <td>${visita.Tipo_Visita}</td>
-        <td>
+      <td>${visita.Pk_Id_Visita_Tecnica}</td>
+      <td>${new Date(visita.Fecha_Visita_Tecnica).toLocaleDateString()}</td>
+      <td>${visita.Tipo_Visita}</td>
+
+      <td>
           <button type="button" class="btn btn-info btn-sm" data-bs-toggle="modal" data-bs-target="#descripcionModal_${
             visita.Pk_Id_Visita_Tecnica
           }">
               Ver Descripción
           </button>
-        </td>
-        <td>${visita.Fk_CC_Usuario}</td>
-        <td>${visita.Fk_Id_Producto}</td>
-        <td>${visita.Fk_Numero_Cuenta_Cliente}</td>
-        <td>
-          <button type="button" class="btn_edit rounded btn btn-primary" data-product-id="${
-            visita.Pk_Id_Visita_Tecnica
-          }"><span></span></button>
-          <button type="button" class="btn_delete rounded btn btn-danger" data-product-id="${
-            visita.Pk_Id_Visita_Tecnica
-          }"><span></span></button>
-        </td>
-      `;
+      </td>
+      <!-- Agrega el modal aquí -->
+      <div class="modal fade" id="descripcionModal_${
+        visita.Pk_Id_Visita_Tecnica
+      }" tabindex="-1" aria-labelledby="descripcionModalLabel_${
+        visita.Pk_Id_Visita_Tecnica
+      }" aria-hidden="true">
+          <div class="modal-dialog modal-dialog-centered">
+              <div class="modal-content">
+                  <div class="modal-header">
+                      <h5 class="modal-title" id="descripcionModalLabel_${
+                        visita.Pk_Id_Visita_Tecnica
+                      }">Descripción de la Visita</h5>
+                      <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
+                  </div>
+                  <div class="modal-body text-wrap">
+                      ${visita.Descripcion_Visita_Tecnica}
+                  </div>
+              </div>
+          </div>
+      </div>
+
+      <td>${visita.Fk_CC_Usuario}</td>
+      <td>${visita.Fk_Id_Producto}</td>
+      <td>${visita.Fk_Numero_Cuenta_Cliente}<td>
+
+      <td>
+            <button type="button" class="btn_edit rounded btn btn-primary" data-visita-id="${
+              visita.Pk_Id_Visita_Tecnica
+            }"><span></span></button>
+            <button type="button" class="btn_delete rounded btn btn-danger" data-product-id="${
+              visita.Pk_Id_Visita_Tecnica
+            }"><span></span></button>
+          </td>
+    `;
       tablaBody_V.appendChild(fila);
 
       fila.querySelector(".btn_delete").addEventListener("click", async (e) => {
@@ -313,13 +207,16 @@ export async function printVisitas() {
       });
     });
 
-    clientes_data.forEach((cliente) => {
+    tecnicos_data.forEach((tecnico) => {
       clienteCC_select_input.appendChild(
         createOption(
-          cliente.Cedula_Cliente,
-          `${cliente.Nombre_Cliente} ${cliente.Apellido_Cliente} CC: ${cliente.Cedula_Cliente}`
+          tecnico.Pk_CC_Usuario,
+          `${tecnico.Nombre_Usuario} ${tecnico.Apellido_Usuario} CC: ${tecnico.Pk_CC_Usuario}`
         )
       );
+    });
+
+    clientes_data.forEach((cliente) => {
       cuentaCliente_select.appendChild(
         createOption(
           cliente.Pk_Numero_Cuenta_Cliente,
